+++ spinetoolbox/config.py
@@ -20,8 +20,9 @@ import sys
 import os
 
 
-REQUIRED_SPINE_ENGINE_VERSION = "0.6.0"
+REQUIRED_SPINE_ENGINE_VERSION = "0.6.1"
 REQUIRED_SPINEDB_API_VERSION = "0.8.5"
+REQUIRED_SPINE_ITEMS_VERSION = "0.1.0"
 LATEST_PROJECT_VERSION = 2
 # SPINE GREEN HTML: #99cc33 RGBa: 153, 204, 51, 255
 # SPINE BLUE HTML: #004ac2 RGBa: 0, 74, 194, 255
@@ -45,11 +46,6 @@ else:
 PLUGINS_PATH = os.path.abspath(os.path.join(_program_root, "plugins"))
 ONLINE_DOCUMENTATION_URL = "https://spine-toolbox.readthedocs.io/en/latest"
 
-# Tool output directory name
-TOOL_OUTPUT_DIR = "output"
-
-# Gimlet default work directory name
-GIMLET_WORK_DIR_NAME = "work"
 
 _on_windows = sys.platform == "win32"