+++ spinetoolbox/project_commands.py
@@ -16,16 +16,7 @@ QUndoCommand subclasses for modifying the project.
 :date:   12.2.2020
 """
 
-from PySide2.QtWidgets import QUndoCommand
-
-
-class SpineToolboxCommand(QUndoCommand):
-    @staticmethod
-    def is_critical():
-        """Returns True if this command needs to be undone before
-        closing the project without saving changes.
-        """
-        return False
+from spine_items.commands import SpineToolboxCommand
 
 
 class SetProjectNameCommand(SpineToolboxCommand):
@@ -237,55 +228,6 @@ class RemoveLinkCommand(SpineToolboxCommand):
 
     def undo(self):
         self.graphics_view._add_link(self.link)
-
-
-class MoveIconCommand(SpineToolboxCommand):
-    def __init__(self, graphics_item):
-        """Command to move icons in the Design view.
-
-        Args:
-            graphics_item (ProjectItemIcon): the icon
-        """
-        super().__init__()
-        self.graphics_item = graphics_item
-        self.previous_pos = {x: x._previous_pos for x in graphics_item.icon_group}
-        self.current_pos = {x: x._current_pos for x in graphics_item.icon_group}
-        if len(graphics_item.icon_group) == 1:
-            self.setText(f"move {list(graphics_item.icon_group)[0]._project_item.name}")
-        else:
-            self.setText("move multiple items")
-
-    def redo(self):
-        for item, current_post in self.current_pos.items():
-            item.setPos(current_post)
-        self.graphics_item.update_links_geometry()
-        self.graphics_item.notify_item_move()
-
-    def undo(self):
-        for item, previous_pos in self.previous_pos.items():
-            item.setPos(previous_pos)
-        self.graphics_item.update_links_geometry()
-        self.graphics_item.notify_item_move()
-
-
-class SetItemSpecificationCommand(SpineToolboxCommand):
-    def __init__(self, item, specification):
-        """Command to set the specification for a Tool.
-
-        Args:
-            item (ProjectItem): the Item
-            specification (ProjectItemSpecification): the new spec
-        """
-        super().__init__()
-        self.item = item
-        self.redo_specification = specification
-        self.setText(f"set specification of {item.name}")
-
-    def redo(self):
-        self.item.do_set_specification(self.redo_specification)
-
-    def undo(self):
-        self.item.undo_set_specification()
 
 
 class AddSpecificationCommand(SpineToolboxCommand):