+++ tests/test_toolboxUI.py
@@ -26,7 +26,8 @@ from PySide2.QtWidgets import QApplication
 from PySide2.QtCore import SIGNAL, Qt, QPoint, QItemSelectionModel
 from PySide2.QtTest import QTest
 from spinetoolbox.project import SpineToolboxProject
-from spinetoolbox.graphics_items import ProjectItemIcon, Link
+from spine_items.graphics_items import ProjectItemIcon
+from spinetoolbox.graphics_items import Link
 from spinetoolbox.project_tree_item import RootProjectTreeItem
 from spinetoolbox.resources_icons_rc import qInitResources
 from .mock_helpers import create_toolboxui, create_project, add_ds, add_dc
@@ -170,7 +171,7 @@ class TestToolboxUI(unittest.TestCase):
         self.assertIsNone(self.toolbox.project())
         with mock.patch("spinetoolbox.ui_main.ToolboxUI.save_project") as mock_save_project, mock.patch(
             "spinetoolbox.project.create_dir"
-        ) as mock_create_dir, mock.patch("spinetoolbox.project_item.create_dir") as mock_create_dir, mock.patch(
+        ) as mock_create_dir, mock.patch("spine_items.project_item.create_dir") as mock_create_dir, mock.patch(
             "spinetoolbox.ui_main.ToolboxUI.update_recent_projects"
         ) as mock_upd_rec_projs:
             self.toolbox.open_project(project_dir)
@@ -704,7 +705,7 @@ class TestToolboxUI(unittest.TestCase):
         self.assertEqual(self.toolbox.project_item_model.n_items(), 1)
         item_index = self.toolbox.project_item_model.find_item("data_connection")
         self.toolbox.ui.treeView_project.selectionModel().select(item_index, QItemSelectionModel.Select)
-        with mock.patch("spinetoolbox.project_item.create_dir") as mock_create_dir:
+        with mock.patch("spine_items.project_item.create_dir") as mock_create_dir:
             self.toolbox.ui.actionCopy.triggered.emit()
             self.toolbox.ui.actionPaste.triggered.emit()
         self.assertEqual(self.toolbox.project_item_model.n_items(), 2)
@@ -717,7 +718,7 @@ class TestToolboxUI(unittest.TestCase):
         self.assertEqual(self.toolbox.project_item_model.n_items(), 1)
         item_index = self.toolbox.project_item_model.find_item("data_connection")
         self.toolbox.ui.treeView_project.selectionModel().select(item_index, QItemSelectionModel.Select)
-        with mock.patch("spinetoolbox.project_item.create_dir") as mock_create_dir:
+        with mock.patch("spine_items.project_item.create_dir") as mock_create_dir:
             self.toolbox.ui.actionDuplicate.triggered.emit()
         self.assertEqual(self.toolbox.project_item_model.n_items(), 2)
         new_item_index = self.toolbox.project_item_model.find_item("data_connection 1")